// mobile/src/shared/types/MultiFormatExtraction.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExtractedImageMetadata } from "./ExtractedImageMetadata";
import type { StructuredBlock } from "./StructuredBlock";
import type { StructuredSection } from "./StructuredSection";

export type MultiFormatExtraction = { document_id: string, source_type: string, title: string, author: string | null, sections: Array<StructuredSection>, blocks: Array<StructuredBlock>, extracted_images: Array<ExtractedImageMetadata>, };